refactored further. roundService.test.js added
